Move reaction serializer to a separate file
