+++ nextjs/components/Message/index.tsx
@@ -11,7 +11,8 @@ import Reactions from './Reactions';
 import Attachments from './Attachments';
 import transform from './utilities/transform';
 import styles from './index.module.css';
-import { SerializedReaction, SerializedAttachment } from 'types/shared';
+import { SerializedAttachment } from 'types/shared';
+import { SerializedReaction } from 'serializers/reaction';
 import parseLinenMessage from 'utilities/message/parsers/linen';
 import parseSlackMessage from 'utilities/message/parsers/slack';
 import parseDiscordMessage from 'utilities/message/parsers/discord';
