+++ nextjs/types/shared.ts
@@ -1,8 +1,3 @@
-export interface SerializedReaction {
-  type: string;
-  count: number;
-}
-
 export interface SerializedAttachment {
   url: string;
   name: string;
